+++ src/main/java/racingcar/dto/request/GameRequestDto.java
@@ -1,4 +1,4 @@
-package racingcar.dto;
+package racingcar.dto.request;
 
 public class GameRequestDto {
     private final String names;
