+++ src/main/java/racingcar/repository/GameRepository.java
@@ -4,7 +4,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import racingcar.dto.GameSaveDto;
+import racingcar.dto.request.GameSaveDto;
 import racingcar.entity.Game;
 
 import javax.sql.DataSource;
