+++ src/main/java/racingcar/repository/PlayerResultRepository.java
@@ -4,7 +4,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import racingcar.dto.PlayerResultSaveDto;
+import racingcar.dto.request.PlayerResultSaveDto;
 import racingcar.entity.PlayerResult;
 
 import javax.sql.DataSource;
