+++ src/main/java/racingcar/service/GameService.java
@@ -3,10 +3,10 @@ package racingcar.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import racingcar.domain.Car;
-import racingcar.dto.GameResponseDto;
-import racingcar.dto.GameResultDto;
-import racingcar.dto.GameSaveDto;
-import racingcar.dto.PlayerResultSaveDto;
+import racingcar.dto.response.GameResponseDto;
+import racingcar.dto.request.GameResultDto;
+import racingcar.dto.request.GameSaveDto;
+import racingcar.dto.request.PlayerResultSaveDto;
 import racingcar.entity.Game;
 import racingcar.entity.PlayerResult;
 import racingcar.repository.GameRepository;
