+++ src/test/java/racingcar/jwp/ApiControllerTest.java
@@ -10,7 +10,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-import racingcar.dto.GameRequestDto;
+import racingcar.dto.request.GameRequestDto;
 
 @SpringBootTest
 @AutoConfigureMockMvc
